refactor: dispatch 의존성 역전
