+++ worker/src/application/sender.ts
@@ -1,0 +1,5 @@
+import { NotificationEntity } from "./dto";
+
+export interface INotificationSender {
+    dispatch(notification: NotificationEntity[]): Promise<void>;
+}
